requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
alembic==1.8.1
anyio==4.2.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
azure-common==1.1.28
azure-core==1.29.7
azure-mgmt-core==1.4.0
azure-mgmt-resource==23.0.1
bcrypt==4.0.1
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.3
click-params==0.3.0
cloudpickle==2.2.1
comm==0.2.1
contourpy==1.2.0
cryptography==41.0.7
cycler==0.12.1
databricks-cli==0.18.0
decorator==5.1.1
distro==1.9.0
docker==6.1.3
entrypoints==0.4
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.99.1
fastapi-utils==0.2.1
Flask==3.0.1
fonttools==4.47.2
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.41
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
httplib2==0.19.1
httptools==0.6.1
idna==3.6
importlib-metadata==7.0.1
ipinfo==5.0.0
ipython==8.20.0
ipywidgets==8.1.1
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jupyterlab-widgets==3.0.9
kiwisolver==1.4.5
Mako==1.3.0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.4
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mlflow==2.9.2
multidict==6.0.4
numpy==1.26.3
oauthlib==3.2.2
orjson==3.8.14
packaging==23.2
pandas==2.2.0
parso==0.8.3
passlib==1.7.4
pexpect==4.9.0
pillow==10.2.0
prompt-toolkit==3.0.43
protobuf==4.25.2
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.2
pycparser==2.21
pydantic==1.10.14
Pygments==2.17.2
PyJWT==2.7.0
PyMySQL==1.0.3
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
querystring-parser==1.2.4
requests==2.31.0
rich==13.7.0
scikit-learn==1.4.0
scipy==1.12.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==1.4.41
SQLAlchemy-Utils==0.38.3
sqlalchemy2-stubs==0.0.2a38
sqlmodel==0.0.8
sqlparse==0.4.4
stack-data==0.6.3
starlette==0.27.0
tabulate==0.9.0
threadpoolctl==3.2.0
traitlets==5.14.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
uvicorn==0.26.0
uvloop==0.19.0
validators==0.18.2
watchfiles==0.21.0
wcwidth==0.2.13
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.0.1
widgetsnbextension==4.0.9
yarl==1.9.4
zenml==0.54.1
zipp==3.17.0
